+++ server/dbi/mysql.js
@@ -97,6 +97,7 @@ const methods = {
     return this._lib.end();
   },
 
+
   // The methods are primitive methods. They perform a simple but usefull
   // operation. Thus, they must be called inside another method as
   // they don't open and close the database.
@@ -104,56 +105,85 @@ const methods = {
   /**
    * Checks if the database is empty.
    *
-   * @method ()
+   * @method (arg1, arg2)
    * @public
-   * @param {}              -,
-   * @returns {Boolean}     returns true if empty otherwise false,
+   * @param {Object}        the database connection,
+   * @param {String}        the name of the database,
+   * @returns {Boolean}     returns true if the db is empty otherwise false,
    * @since 0.0.0
    */
-  async isDbEmpty() {
-    return true;
+  async isDbEmpty(cn, db) {
+    const SQL = 'SELECT COUNT(DISTINCT `table_name`) AS TotalNumberOfTablesFROM `information_schema`.`columns` WHERE `table_schema` = ?';
+    const resp = await this._lib.query(cn, SQL, [db]);
+    return resp[0].TotalNumberOfTables === 0;
   },
 
   /**
    * Checks if the table exists.
-   * (must be be overwritten)
    *
-   * @method (arg1)
+   * @method (arg1, arg2)
    * @public
+   * @param {Object}        the database connection,
    * @param {String}        the name of the table,
    * @returns {Boolean}     returns true if the table exists otherwise false,
    * @since 0.0.0
    */
-  async isTable(table) {
-    return table;
+  async isTable(cn, table) {
+    const SQL = `SELECT * FROM information_schema.tables
+      WHERE table_schema = '${this._db}' AND table_name = '${table}'`;
+    const resp = await this._lib.query(cn, SQL);
+    return resp.length > 0;
   },
 
   /**
    * Returns the table structure.
-   * (must be be overwritten)
    *
-   * @method (arg1)
+   * @method (arg1, arg2)
    * @public
+   * @param {Object}        the database connection,
    * @param {String}        the name of the table,
    * @returns {Array}       returns the structure of the table,
    * @since 0.0.0
    */
-  async getTableStructure(table) {
-    return table;
+  async getTableStructure(cn, table) {
+    const r = await this._lib.query(cn, `SHOW COLUMNS FROM ${table}`);
+
+    // Standardize:
+    const struct = [];
+    for (let i = 0; i < r.length; i++) {
+      let def = 'NULL';
+      if (r[i].Null === 'NO' && r[i].Default === null) {
+        def = 'None';
+      } else if (r[i].Null === 'NO') {
+        def = r[i].Default;
+      }
+      struct.push({
+        column: i,
+        name: r[i].Field,
+        type: r[i].Type.toUpperCase(),
+        notNULL: r[i].Null === 'NO' ? 'Yes' : 'No',
+        default: def,
+        key: r[i].Key === 'PRI' ? 'primary' : '',
+        extra: r[i].Extra,
+      });
+    }
+
+    return struct;
   },
 
   /**
    * Checks if the table is empty.
-   * (must be be overwritten)
    *
-   * @method ()
+   * @method (arg1, arg2)
    * @public
-   * @param {}              -,
-   * @returns {Boolean}     returns true if empty otherwis false,
+   * @param {Object}        the database connection,
+   * @param {String}        the name of the table,
+   * @returns {Boolean}     returns true if empty otherwise false,
    * @since 0.0.0
    */
-  async isTableEmpty(table) {
-    return table;
+  async isTableEmpty(cn, table) {
+    const resp = await this._lib.query(cn, `select count(*) from ${table}`);
+    return resp[0]['count(*)'] === 0;
   },
 };